+++ HISTORY.rst
@@ -9,6 +9,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * BSSLX-102: Migrate the mrp order report in V11
+* BSSLX-123: Add "Include in statistics" custom filter in V11 `report.all.channels.sales` report
 * BSSLX-127: Uninstall account_bank_statement_import_camt.
   Install account_bank_statement_import_camt_oca.